Add more unit tests and speed up the security job in pr.yaml to use the build cache
