+++ .github/workflows/integration.yaml
@@ -84,7 +84,7 @@ jobs:
       - name: Install CRDs
         run: |
           kubectl create namespace ${NAMESPACE} || true
-          kubectl apply -f deploy/crds/
+          kubectl apply -k deploy/crds/
 
       - name: Install RBAC
         run: kubectl apply -f deploy/rbac/
