+++ deploy/crds/arecords.crd.yaml
@@ -34,7 +34,6 @@ spec:
                   description: Record name (@ for apex)
                 ipv4Address:
                   type: string
-                  pattern: '^(\\d{1,3}\\.){3}\\d{1,3}$'
                   description: IPv4 address
                 ttl:
                   type: integer
