+++ deploy/crds/kustomization.yaml
@@ -1,0 +1,14 @@
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
+
+resources:
+  - aaaarecords.crd.yaml
+  - arecords.crd.yaml
+  - bind9instances.crd.yaml
+  - caarecords.crd.yaml
+  - cnamerecords.crd.yaml
+  - dnszones.crd.yaml
+  - mxrecords.crd.yaml
+  - nsrecords.crd.yaml
+  - srvrecords.crd.yaml
+  - txtrecords.crd.yaml
