+++ docs/src/operations/faq.md
@@ -137,7 +137,7 @@ kubectl delete arecords,aaaarecords,cnamerecords,mxrecords,txtrecords \
 
 ### How do I upgrade Bindy?
 
-1. Update CRDs: `kubectl apply -f deploy/crds/`
+1. Update CRDs: `kubectl apply -k deploy/crds/`
 2. Update controller: `kubectl set image deployment/bindy controller=new-image`
 3. Monitor rollout: `kubectl rollout status deployment/bindy -n dns-system`
