+++ src/lib.rs
@@ -64,7 +64,14 @@
 pub mod bind9;
 pub mod bind9_resources;
 pub mod crd;
+pub mod crd_docs;
 pub mod reconcilers;
 
+#[cfg(test)]
+mod bind9_resources_tests;
+#[cfg(test)]
+mod bind9_tests;
+#[cfg(test)]
+mod crd_docs_tests;
 #[cfg(test)]
 mod crd_tests;
